Fix voxel size reading in SubSamplingImageAlgorithm
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@63908 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/public/resampling/subsamplingimagealgorithm.h
@@ -157,7 +157,7 @@ namespace aims {
       //           << carto::toString(dim[2]) << ", "
       //           << carto::toString(dim[3]) << "]"
       //           << std::endl;
-      std::vector<double> vsv( 4, 1. );
+      std::vector<float> vsv( 4, 1. );
       in->header().getProperty( "voxel_size", vsv );
       vsv[0] *= _win_size_x;
       vsv[1] *= _win_size_y;
